Upgrade C++ bindings for the latest libscran libraries. (#87)
This uses the latest versions of assorthead, knncolle and mattress. This uses a
new MNN algorithm, adds quantile options for marker summarization, adds
quantile blocking options for marker and variance estimates, and adds blocking
for scaling of modalities. We also update to igraph 1.0.0 for graph clustering.

In the bindings themselves, we use sanisizer to avoid overflow when defining
pybind11 containers. We also use pybind11's forcecast to enforce the format of
the input arrays while supporting different user-provided arguments.